move DID doc code out to separate file
